Pull getAllFoods & getAllFoodsValidation into serviceFoods.js
