+++ lib/index.js
@@ -1,6 +1,6 @@
 import { appendMeals, getEachFoodDiary, filterFoods, removeMealFood } from './domIndex'
 import { validateForm, getEachFood, removeFood, removeValidations, createInputField, removeInputField } from './domFoods'
-import { postFood } from './serviceFoods'
+import { postFood, getAllFoodsValidation } from './serviceFoods'
 
 export const handleResponse = (response) => {
   return response.json()
@@ -18,20 +18,20 @@ export const handleResponse = (response) => {
   })
 }
 
-export const getAllFoods = (meals) => {
-  $('#all-foods').html('<tr><th>Food</th><th>Calories</th></tr>')
-  fetch('https://quantifiedself-backend.herokuapp.com/api/v1/foods')
-    .then((response) => handleResponse(response))
-    .then((foods) => getEachFood(meals, foods))
-    .catch((error) => console.error({error}))
-}
-
-export const getAllFoodsValidation = () => {
-  fetch(`https://quantifiedself-backend.herokuapp.com/api/v1/meals`)
-    .then((response) => handleResponse(response))
-    .then((meals) => getAllFoods(meals))
-    .catch((error) => console.error({error}))
-}
+// export const getAllFoods = (meals) => {
+//   $('#all-foods').html('<tr><th>Food</th><th>Calories</th></tr>')
+//   fetch('https://quantifiedself-backend.herokuapp.com/api/v1/foods')
+//     .then((response) => handleResponse(response))
+//     .then((foods) => getEachFood(meals, foods))
+//     .catch((error) => console.error({error}))
+// }
+//
+// export const getAllFoodsValidation = () => {
+//   fetch(`https://quantifiedself-backend.herokuapp.com/api/v1/meals`)
+//     .then((response) => handleResponse(response))
+//     .then((meals) => getAllFoods(meals))
+//     .catch((error) => console.error({error}))
+// }
 
 const getAllFoodsDiary = () => {
   $('#all-foods-diary').html('<tr><th>Food</th><th class=sort>Calories</th></tr>')
